Move the logIn method to the auth service
