+++ photo-sharing-application-front-end/src/app/services/auth/auth.service.ts
@@ -1,7 +1,9 @@
 import { Injectable, Signal, signal } from '@angular/core';
 import { Account } from '../../models/account';
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { environment } from 'src/environments/environment';
+import { Credentials } from 'src/app/models/credentials';
+import { Observable } from 'rxjs';
 
 @Injectable({
   providedIn: 'root',
@@ -20,5 +22,22 @@ export class AuthService {
       next: (d) => this.#accountDetails.set(d.data),
       error: () => this.#accountDetails.set(undefined),
     });
+  }
+
+  logIn(credentials: Credentials): Observable<any> {
+    const httpOptions = {
+      headers: new HttpHeaders({
+        Authorization:
+          'Basic ' +
+          Buffer.from(`${credentials.email}:${credentials.password}`).toString(
+            'base64'
+          ),
+      }),
+    };
+    return this.http.post(
+      `${environment.apiUrl}/api/v1/sessions`,
+      {},
+      httpOptions
+    );
   }
 }
