+++ photo-sharing-application-front-end/src/app/services/login/login.service.ts
@@ -12,20 +12,5 @@ import { environment } from 'src/environments/environment';
 export class LoginService {
   constructor(private http: HttpClient) {}
 
-  logIn(credentials: Credentials): Observable<any> {
-    const httpOptions = {
-      headers: new HttpHeaders({
-        Authorization:
-          'Basic ' +
-          Buffer.from(`${credentials.email}:${credentials.password}`).toString(
-            'base64'
-          ),
-      }),
-    };
-    return this.http.post(
-      `${environment.apiUrl}/api/v1/sessions`,
-      {},
-      httpOptions
-    );
-  }
+  
 }
